Struct carambolage::game::car::Car

pub(crate) struct Car {
    pub(crate) position: Vector3<f32>,
    pub(crate) rotation: Vector3<f32>,
    _velocity: Vector3<f32>,
    _force: Vector3<f32>,
    _mass: f32,
    boost: f32,
    pub(crate) model: Model,
    pub(crate) cuboid: Cuboid<f32>,
}

A GameObject controlled by a player.

Fields
position: Vector3<f32> rotation: Vector3<f32> _velocity: Vector3<f32> _force: Vector3<f32> _mass: f32 boost: f32 model: Model cuboid: Cuboid<f32>
Methods

impl Car

pub fn new(
model: &str,
color_palette: &str,
position: Vector3<f32>,
mass: f32
) -> Car

Create a new Car.

For model and color_palette see model_from_id() and color_from_id(). mass is the mass of the car in [kg]

pub(super) fn update(
&mut self,
dt: f32,
controller: Option<Controller>
)

Update the car position and velocity based on the internal car state for a given time step.

pub(super) fn draw(
&self,
view: &Matrix4<f32>,
projection: &Matrix4<f32>
)

Draw the car to the currently bound framebuffer.

pub fn model_from_id(id: u32) -> String

Return a Car file name from an id.

1:kart 2:parsche 3:farara 4:lamba 5:gtc1 6:gtc2 7:formula

pub fn color_from_id(id: u32) -> String

Return a Car color palette file name from an id.

1:blue 2:green 3:lime 4:orange 5:purple 6:red 7:yellow