Struct carambolage::game::transform::Transform

pub struct Transform {
    needs_update: bool,
    position: Vector3<f32>,
    rotation: Vector3<f32>,
    scale: Vector3<f32>,
    matrix: Matrix4<f32>,
}

Position, rotation and scale of an object.

Every gameobject in a scene has (should have) a Transform. Its purpose is to store and manipulate the position, rotation and scale of the gameobject and determine its model matrix for rendering.

Fields
needs_update: bool position: Vector3<f32> rotation: Vector3<f32> scale: Vector3<f32> matrix: Matrix4<f32>
Methods

impl Transform

pub fn new() -> Transform

pub fn set_position(&mut self, position: Vector3<f32>)

Sets the transforms position in world coordinates.

pub fn set_rotation(&mut self, rotation: Vector3<f32>)

Sets the rotation of eulerAngles in the following order: (I hope the order is right and it is in degrees not radians)

z degrees around the z axis

x degrees around the x axis

y degrees around the y axis

pub fn set_scale(&mut self, scale: Vector3<f32>)

Sets the scale of the object for each axis.

pub fn position(&self) -> Vector3<f32>

Sets the objects origin relative to world coordinates.

pub fn rotation(&self) -> Vector3<f32>

Returns the rotation of eulerAngles. x: yaw y: pitch z: roll

pub fn scale(&self) -> Vector3<f32>

Returns the scale of the object for each axis.

pub fn matrix(&mut self) -> Matrix4<f32>

Returns the model matrix according to scale, rotation and position of the object in world coordinates. Note: The object has to be mutable because it stores a model matrix, which will be updated in this function if some parameter has changed since the last matrix request.